Transition userIds to Long data type due to Telegram announcement

+++ core/src/com/bot4s/telegram/Implicits.scala
@@ -20,7 +20,7 @@ object Implicits {
 
     def altWithUrl(url: String) = s"[$s]($url)"
 
-    def mention(userId: Int) = s"[$s](tg://user?id=$userId)"
+    def mention(userId: Long) = s"[$s](tg://user?id=$userId)"
 
     def inlineCode = s"`$s`"
